requirements.txt
amqp==5.0.6
asgiref==3.4.1
astroid==2.5.1
billiard==3.6.4.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cycler==0.10.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.5
django-cors-headers==3.8.0
django-filter==2.4.0
django-import-export==2.5.0
django-wkhtmltopdf==3.3.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
et-xmlfile==1.1.0
gunicorn==20.1.0
isort==5.7.0
kiwisolver==1.3.1
kombu==5.1.0
lazy-object-proxy==1.5.2
lml==0.1.0
MarkupPy==1.14
matplotlib==3.3.4
mccabe==0.6.1
numpy==1.20.1
odfpy==1.4.1
openpyxl==3.0.8
pandas==1.2.3
pickle4==0.0.1
Pillow==8.1.2
prompt-toolkit==3.0.20
psycopg2==2.9.1
psycopg2-binary==2.9.1
pyexcel-io==0.6.4
pyexcel-xls==0.6.2
pyexcel-xlsx==0.6.0
PyJWT==2.0.0
pylint==2.7.2
pyparsing==2.4.7
pypi==2.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
qrcode==7.2
redis==3.5.3
scipy==1.6.1
seaborn==0.11.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
